refactor: split notes.test.ts and rename test files to match sources
Split the monolithic notes.test.ts (1532 lines) into individual test
files per source module. Renamed derive/hash/seed test files to match
their renamed source counterparts (derivation/encoding/mnemonic).

New test files: token-utils, note, shield-note, transact-note,
unshield-note. Decrypt-commitment tests merged into commitment.test.ts,
memo roundtrip test merged into memo.test.ts.
